refactor CamelCase
