+++ lib/main.dart
@@ -11,7 +11,7 @@ import 'package:flutter/material.dart';
 import 'package:Videotheque/views/app_view.dart';
 import 'package:Videotheque/views/movie_view/movie_view.dart';
 import 'package:Videotheque/views/person_view/person_view.dart';
-import 'package:Videotheque/views/search_view/SearchView.dart';
+import 'package:Videotheque/views/SearchView/SearchView.dart';
 import 'package:Videotheque/views/collection_view/collection_view.dart';
 import 'package:Videotheque/views/tv_view/tv_view.dart';
 import 'package:Videotheque/views/authView/AuthView.dart';
