+++ lib/views/SearchView/ResultView.dart
