+++ lib/views/SearchView/components/CollectionCardComponent.dart
