+++ lib/views/SearchView/components/IconError.dart
