+++ lib/views/SearchView/components/KnownForMovieComponent.dart
