+++ lib/views/SearchView/components/MovieCardComponent.dart
