+++ lib/views/SearchView/components/TvCardComponent.dart
