+++ lib/views/app_view.dart
@@ -1,7 +1,7 @@
 import 'package:Videotheque/views/account_view/account_view.dart';
 import 'package:Videotheque/views/home_view/HomeView.dart';
 import 'package:Videotheque/views/library_view/library_view.dart';
-import 'package:Videotheque/views/search_view/SearchView.dart';
+import 'package:Videotheque/views/SearchView/SearchView.dart';
 import 'package:bottom_navy_bar/bottom_navy_bar.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
